+++ src/RhinoInside.Revit.GH/Components/Documents/DesignOptions/QueryDesignOptions.cs
@@ -7,6 +7,10 @@ using ARDB = Autodesk.Revit.DB;
 
 namespace RhinoInside.Revit.GH.Components.DesignOptions
 {
+  using External.DB;
+  using External.DB.Extensions;
+
+  [ComponentVersion(introduced: "1.0", updated: "1.36")]
   public class QueryDesignOptionSets : ElementCollectorComponent
   {
     public override Guid ComponentGuid => new Guid("B31E7605-87F6-421B-84C5-8B00BFD0BC88");
@@ -36,7 +40,7 @@ namespace RhinoInside.Revit.GH.Components.DesignOptions
     protected override ParamDefinition[] Inputs => inputs;
     static readonly ParamDefinition[] inputs =
     {
-      new ParamDefinition(new Parameters.Document(), ParamRelevance.Occasional),
+      new ParamDefinition(new Parameters.ModelInstance(), ParamRelevance.Occasional),
       ParamDefinition.Create<Param_String>("Name", "N", "Design Option name", GH_ParamAccess.item, optional: true),
       ParamDefinition.Create<Parameters.ElementFilter>("Filter", "F", "Filter", GH_ParamAccess.item, optional: true, relevance: ParamRelevance.Occasional),
     };
@@ -44,23 +48,21 @@ namespace RhinoInside.Revit.GH.Components.DesignOptions
     protected override ParamDefinition[] Outputs => outputs;
     static readonly ParamDefinition[] outputs =
     {
-      ParamDefinition.Create<Parameters.Element>("Design Option Sets", "V", "Design Option Sets list", GH_ParamAccess.list)
+      ParamDefinition.Create<Parameters.Element>("Design Option Sets", "S", "Design Option Sets list", GH_ParamAccess.list)
     };
 
     protected override void TrySolveInstance(IGH_DataAccess DA)
     {
-      if (!Parameters.Document.GetDataOrDefault(this, DA, "Document", out var doc))
-        return;
-
+      if (!Parameters.ModelInstance.TryGetOrCurrent(this, DA, "Model", out var model)) return;
       if (!Params.TryGetData(DA, "Name", out string name)) return;
       if (!Params.TryGetData(DA, "Filter", out ARDB.ElementFilter filter, x => x.IsValidObject)) return;
 
-      using (var collector = new ARDB.FilteredElementCollector(doc))
+      using (var collector = new ARDB.FilteredElementCollector(model.ModelDocument.Value))
       {
         var optionsCollector = collector.WherePasses(ElementFilter);
 
         if (filter is object)
-          optionsCollector = optionsCollector.WherePasses(filter);
+          optionsCollector = optionsCollector.WherePasses(filter, model.ModelInstance.Value);
 
         if (name is object && TryGetFilterStringParam(ARDB.BuiltInParameter.OPTION_SET_NAME, ref name, out var nameFilter))
           optionsCollector = optionsCollector.WherePasses(nameFilter);
@@ -75,12 +77,14 @@ namespace RhinoInside.Revit.GH.Components.DesignOptions
           "Design Option Sets",
           options.
           Select(x => new Types.DesignOptionSet(x)).
+          AtModel(model).
           TakeWhileIsNotEscapeKeyDown(this)
         );
       }
     }
   }
 
+  [ComponentVersion(introduced: "1.0", updated: "1.36")]
   public class QueryDesignOptions : ElementCollectorComponent
   {
     public override Guid ComponentGuid => new Guid("6804582B-E6FD-4825-AA71-B59346F149CD");
@@ -111,8 +115,8 @@ namespace RhinoInside.Revit.GH.Components.DesignOptions
     protected override ParamDefinition[] Inputs => inputs;
     static readonly ParamDefinition[] inputs =
     {
-      new ParamDefinition(new Parameters.Document(), ParamRelevance.Occasional),
-      ParamDefinition.Create<Parameters.Element>("Design Option Set", "DOS", string.Empty, GH_ParamAccess.item, optional: true),
+      new ParamDefinition(new Parameters.ModelInstance(), ParamRelevance.Occasional),
+      ParamDefinition.Create<Parameters.Element>("Design Option Set", "DOS", string.Empty, GH_ParamAccess.item, optional:true),
       ParamDefinition.Create<Param_String>("Name", "N", "Design Option name", GH_ParamAccess.item, optional: true),
       ParamDefinition.Create<Param_Boolean>("Primary", "P", "Design Option is primary", GH_ParamAccess.item, optional: true),
       ParamDefinition.Create<Parameters.ElementFilter>("Filter", "F", "Filter", GH_ParamAccess.item, optional: true, relevance: ParamRelevance.Occasional),
@@ -121,60 +125,59 @@ namespace RhinoInside.Revit.GH.Components.DesignOptions
     protected override ParamDefinition[] Outputs => outputs;
     static readonly ParamDefinition[] outputs =
     {
-      ParamDefinition.Create<Parameters.Element>("Design Options", "V", "Design Options list", GH_ParamAccess.list)
+      ParamDefinition.Create<Parameters.Element>("Design Options", "O", "Design Options list", GH_ParamAccess.list)
     };
 
     protected override void TrySolveInstance(IGH_DataAccess DA)
     {
-      if (!Parameters.Document.GetDataOrDefault(this, DA, "Document", out var doc))
-        return;
-
+      if (!Parameters.ModelInstance.TryGetOrCurrent(this, DA, "Model", out var model)) return;
       if (!Params.TryGetData(DA, "Design Option Set", out Types.DesignOptionSet set)) return;
       if (!Params.TryGetData(DA, "Name", out string name)) return;
       if (!Params.TryGetData(DA, "Primary", out bool? primary)) return;
       if (!Params.TryGetData(DA, "Filter", out ARDB.ElementFilter filter, x => x.IsValidObject)) return;
 
-      if (!(set?.Document is null || doc.Equals(set.Document)))
-        throw new System.ArgumentException("Wrong Document.", "Design Option Set");
-
-      using (var collector = new ARDB.FilteredElementCollector(doc))
+      if (set?.AssertValidModel(model) != false)
       {
-        var optionsCollector = collector.WherePasses(ElementFilter);
-
-        if (filter is object)
-          optionsCollector = optionsCollector.WherePasses(filter);
-
-        if (set is object && TryGetFilterElementIdParam(ARDB.BuiltInParameter.OPTION_SET_ID, set.Id, out var optionSetFilter))
-          optionsCollector = optionsCollector.WherePasses(optionSetFilter);
-
-        if (name is object && TryGetFilterStringParam(ARDB.BuiltInParameter.OPTION_NAME, ref name, out var nameFilter))
-          optionsCollector = optionsCollector.WherePasses(nameFilter);
-
-        var options = collector.Cast<ARDB.DesignOption>();
-
-        if (primary.HasValue)
-          options = options.Where(x => x.IsPrimary == primary.Value);
-
-        if (name is object)
-          options = options.Where(x => x.get_Parameter(ARDB.BuiltInParameter.OPTION_NAME).AsString().IsSymbolNameLike(name));
-
-        if
-        (
-          (set is null || set.Id == ARDB.ElementId.InvalidElementId) &&
-          (name is null || name == "Main Model") &&
-          primary != false
-        )
+        using (var collector = new ARDB.FilteredElementCollector(model.ModelDocument.Value))
         {
-          options = Enumerable.Repeat(default(ARDB.DesignOption), 1).Concat(options);
-        }
+          var optionsCollector = collector.WherePasses(ElementFilter);
 
-        DA.SetDataList
-        (
-          "Design Options",
-          options.
-          Select(x => new Types.DesignOption(x)).
-          TakeWhileIsNotEscapeKeyDown(this)
-        );
+          if (filter is object)
+            optionsCollector = optionsCollector.WherePasses(filter, model.ModelInstance.Value);
+
+          if (set is object && TryGetFilterElementIdParam(ARDB.BuiltInParameter.OPTION_SET_ID, set.Id, out var optionSetFilter))
+            optionsCollector = optionsCollector.WherePasses(optionSetFilter);
+
+          if (name is object && TryGetFilterStringParam(ARDB.BuiltInParameter.OPTION_NAME, ref name, out var nameFilter))
+            optionsCollector = optionsCollector.WherePasses(nameFilter);
+
+          var options = collector.Cast<ARDB.DesignOption>();
+
+          if (primary.HasValue)
+            options = options.Where(x => x.IsPrimary == primary.Value);
+
+          if (name is object)
+            options = options.Where(x => x.get_Parameter(ARDB.BuiltInParameter.OPTION_NAME).AsString().IsSymbolNameLike(name));
+
+          if
+          (
+            (set is null || set.Id == ARDB.ElementId.InvalidElementId) &&
+            (string.IsNullOrEmpty(name)) &&
+            primary != false
+          )
+          {
+            options = Enumerable.Repeat(default(ARDB.DesignOption), 1).Concat(options);
+          }
+
+          DA.SetDataList
+          (
+            "Design Options",
+            options.
+            Select(x => new Types.DesignOption(x)).
+            AtModel(model).
+            TakeWhileIsNotEscapeKeyDown(this)
+          );
+        }
       }
     }
   }