+++ src/RhinoInside.Revit.GH/Components/Documents/Worksets/QueryWorksets.cs
@@ -8,7 +8,7 @@ using ARDB = Autodesk.Revit.DB;
 
 namespace RhinoInside.Revit.GH.Components.Worksets
 {
-  [ComponentVersion(introduced: "1.2")]
+  [ComponentVersion(introduced: "1.2", updated:"1.36")]
   public class QueryWorksets : ElementCollectorComponent
   {
     public override Guid ComponentGuid => new Guid("311316BA-81C7-495C-8A20-B7974091D6B1");
@@ -44,7 +44,7 @@ namespace RhinoInside.Revit.GH.Components.Worksets
     protected override ParamDefinition[] Inputs => inputs;
     static readonly ParamDefinition[] inputs =
     {
-      new ParamDefinition(new Parameters.Document(), ParamRelevance.Occasional),
+      new ParamDefinition(new Parameters.ModelInstance(), ParamRelevance.Occasional),
       ParamDefinition.Create<Parameters.Param_Enum<Types.WorksetKind>>
         ("Kind", "K", "Workset kind", defaultValue: ARDB.WorksetKind.UserWorkset, optional: true),
       ParamDefinition.Create<Param_String>
@@ -58,11 +58,11 @@ namespace RhinoInside.Revit.GH.Components.Worksets
 
     protected override void TrySolveInstance(IGH_DataAccess DA)
     {
-      if (!Parameters.Document.GetDataOrDefault(this, DA, "Document", out var doc)) return;
+      if (!Parameters.ModelInstance.TryGetOrCurrent(this, DA, "Model", out var model)) return;
       if (!Params.TryGetData(DA, "Kind", out Types.WorksetKind kind)) return;
       if (!Params.TryGetData(DA, "Name", out string name)) return;
 
-      using (var collector = new ARDB.FilteredWorksetCollector(doc))
+      using (var collector = new ARDB.FilteredWorksetCollector(model.ModelDocument.Value))
       {
         var worksetCollector = collector;
 
@@ -77,7 +77,7 @@ namespace RhinoInside.Revit.GH.Components.Worksets
         DA.SetDataList
         (
           "Worksets",
-          worksets.Select(x => new Types.Workset(doc, x)).
+          worksets.Select(x => new Types.Workset(model.ModelDocument.Value, x)).
           TakeWhileIsNotEscapeKeyDown(this)
         );
       }