+++ src/RhinoInside.Revit.GH/Components/Element/Assembly/QueryAssemblies.cs
@@ -6,7 +6,9 @@ using ARDB = Autodesk.Revit.DB;
 
 namespace RhinoInside.Revit.GH.Components.Assemblies
 {
-  [ComponentVersion(introduced: "1.2", updated: "1.5")]
+  using External.DB;
+
+  [ComponentVersion(introduced: "1.2", updated: "1.36")]
   public class QueryAssemblies : ElementCollectorComponent
   {
     public override Guid ComponentGuid => new Guid("fd5b45c3-7f55-4ad8-abbe-e871f95b4988");
@@ -26,7 +28,7 @@ namespace RhinoInside.Revit.GH.Components.Assemblies
     protected override ParamDefinition[] Inputs => inputs;
     static readonly ParamDefinition[] inputs =
     {
-      new ParamDefinition(new Parameters.Document(), ParamRelevance.Occasional),
+      new ParamDefinition(new Parameters.ModelInstance(), ParamRelevance.Occasional),
       ParamDefinition.Create<Param_String>("Type Name", "N", "Assembly type name", GH_ParamAccess.item, optional: true),
       ParamDefinition.Create<Parameters.ElementFilter>("Filter", "F", "Filter", GH_ParamAccess.item, optional: true),
     };
@@ -39,8 +41,7 @@ namespace RhinoInside.Revit.GH.Components.Assemblies
 
     protected override void TrySolveInstance(IGH_DataAccess DA)
     {
-      if (!Parameters.Document.GetDataOrDefault(this, DA, "Document", out var doc))
-        return;
+      if (!Parameters.ModelInstance.TryGetOrCurrent(this, DA, "Model", out var model)) return;
 
       string name = null;
       DA.GetData("Type Name", ref name);
@@ -48,12 +49,12 @@ namespace RhinoInside.Revit.GH.Components.Assemblies
       ARDB.ElementFilter filter = null;
       DA.GetData("Filter", ref filter);
 
-      using (var collector = new ARDB.FilteredElementCollector(doc))
+      using (var collector = new ARDB.FilteredElementCollector(model.ModelDocument.Value))
       {
         var assemblyCollector = collector.WherePasses(ElementFilter);
 
         if (filter is object)
-          assemblyCollector = assemblyCollector.WherePasses(filter);
+          assemblyCollector = assemblyCollector.WherePasses(filter, model.ModelInstance.Value);
 
         if (TryGetFilterStringParam(ARDB.BuiltInParameter.ALL_MODEL_TYPE_NAME, ref name, out var assemblyNameFilter))
           assemblyCollector = assemblyCollector.WherePasses(assemblyNameFilter);
@@ -68,6 +69,7 @@ namespace RhinoInside.Revit.GH.Components.Assemblies
           "Assemblies",
           assemblies.
           Select(x => new Types.AssemblyInstance(x)).
+          AtModel(model).
           TakeWhileIsNotEscapeKeyDown(this)
         );
       }