+++ src/RhinoInside.Revit.GH/Components/Topology/QuerySpatialElements.cs
@@ -19,16 +19,7 @@ namespace RhinoInside.Revit.GH.Components.Topology
     protected override ParamDefinition[] Inputs => inputs;
     static readonly ParamDefinition[] inputs =
     {
-      new ParamDefinition
-      (
-        new Parameters.Document()
-        {
-          Name = "Document",
-          NickName = "DOC",
-          Description = "Document",
-          Optional = true
-        }, ParamRelevance.Occasional
-      ),
+      new ParamDefinition(new Parameters.ModelInstance(), ParamRelevance.Occasional),
       new ParamDefinition
       (
         new Param_Point()
@@ -112,7 +103,7 @@ namespace RhinoInside.Revit.GH.Components.Topology
     };
   }
 
-  [ComponentVersion(introduced: "1.7", updated: "1.12")]
+  [ComponentVersion(introduced: "1.7", updated: "1.36")]
   public class QueryAreas : ElementCollectorComponent
   {
     public override Guid ComponentGuid => new Guid("D1940EB3-B81B-4E57-8F5A-94D045BFB509");
@@ -134,16 +125,7 @@ namespace RhinoInside.Revit.GH.Components.Topology
     protected override ParamDefinition[] Inputs => inputs;
     static readonly ParamDefinition[] inputs =
     {
-      new ParamDefinition
-      (
-        new Parameters.Document()
-        {
-          Name = "Document",
-          NickName = "DOC",
-          Description = "Document",
-          Optional = true
-        }, ParamRelevance.Occasional
-      ),
+      new ParamDefinition(new Parameters.ModelInstance(), ParamRelevance.Occasional),
       new ParamDefinition
       (
         new Param_Point()
@@ -243,7 +225,7 @@ namespace RhinoInside.Revit.GH.Components.Topology
 
     protected override void TrySolveInstance(IGH_DataAccess DA)
     {
-      if (!Parameters.Document.TryGetDocumentOrCurrent(this, DA, "Document", out var doc)) return;
+      if (!Params.TryGetData(DA, "Model", out Types.IGH_ModelInstance model, x => x.IsValid)) return;
       if (!Params.TryGetData(DA, "Point", out Point3d? point)) return;
       if (!Params.TryGetData(DA, "Placed", out bool? placed)) return;
       if (!Params.TryGetData(DA, "Number", out string number)) return;
@@ -260,12 +242,12 @@ namespace RhinoInside.Revit.GH.Components.Topology
       }
 
       var tol = GeometryTolerance.Model;
-      using (var collector = new ARDB.FilteredElementCollector(doc.Value))
+      using (var collector = new ARDB.FilteredElementCollector(model.ModelDocument.Value))
       {
         var elementsCollector = collector.WherePasses(ElementFilter);
 
         if (filter is object)
-          elementsCollector = elementsCollector.WherePasses(filter);
+          elementsCollector = elementsCollector.WherePasses(filter, model.ModelInstance.Value);
 
         if (TryGetFilterStringParam(ARDB.BuiltInParameter.ROOM_NUMBER, ref number, out var numberFilter))
           elementsCollector = elementsCollector.WherePasses(numberFilter);
@@ -314,12 +296,12 @@ namespace RhinoInside.Revit.GH.Components.Topology
           );
         }
 
-        DA.SetDataList("Areas", areas.TakeWhileIsNotEscapeKeyDown(this));
+        DA.SetDataList("Areas", areas.AtModel(model).TakeWhileIsNotEscapeKeyDown(this));
       }
     }
   }
 
-  [ComponentVersion(introduced: "1.12")]
+  [ComponentVersion(introduced: "1.12", updated: "1.36")]
   public class QueryAreaSchemes : ElementCollectorComponent
   {
     public override Guid ComponentGuid => new Guid("3E2A753B-6321-4C4C-B697-022D1F04F213");
@@ -332,7 +314,7 @@ namespace RhinoInside.Revit.GH.Components.Topology
     (
       name: "Query Area Schemes",
       nickname: "AreaSchemes",
-      description: "Get document area shemes list",
+      description: "Get document area schemes list",
       category: "Revit",
       subCategory: "Topology"
     )
@@ -341,16 +323,7 @@ namespace RhinoInside.Revit.GH.Components.Topology
     protected override ParamDefinition[] Inputs => inputs;
     static readonly ParamDefinition[] inputs =
     {
-      new ParamDefinition
-      (
-        new Parameters.Document()
-        {
-          Name = "Document",
-          NickName = "DOC",
-          Description = "Document",
-          Optional = true
-        }, ParamRelevance.Occasional
-      ),
+      new ParamDefinition(new Parameters.ModelInstance(), ParamRelevance.Occasional),
       new ParamDefinition
       (
         new Param_String()
@@ -390,12 +363,12 @@ namespace RhinoInside.Revit.GH.Components.Topology
 
     protected override void TrySolveInstance(IGH_DataAccess DA)
     {
-      if (!Parameters.Document.TryGetDocumentOrCurrent(this, DA, "Document", out var doc)) return;
+      if (!Params.TryGetData(DA, "Model", out Types.IGH_ModelInstance model, x => x.IsValid)) return;
       if (!Params.TryGetData(DA, "Name", out string name)) return;
       if (!Params.TryGetData(DA, "Gross Building", out bool? gross)) return;
 
       var tol = GeometryTolerance.Model;
-      using (var collector = new ARDB.FilteredElementCollector(doc.Value))
+      using (var collector = new ARDB.FilteredElementCollector(model.ModelDocument.Value))
       {
         var elementsCollector = collector.WherePasses(ElementFilter);
 
@@ -410,12 +383,12 @@ namespace RhinoInside.Revit.GH.Components.Topology
         if (!string.IsNullOrEmpty(name))
           areas = areas.Where(x => x.Nomen.IsSymbolNameLike(name));
 
-        DA.SetDataList("Area Schemes", areas.TakeWhileIsNotEscapeKeyDown(this));
+        DA.SetDataList("Area Schemes", areas.AtModel(model).TakeWhileIsNotEscapeKeyDown(this));
       }
     }
   }
 
-  [ComponentVersion(introduced: "1.7", updated: "1.9")]
+  [ComponentVersion(introduced: "1.7", updated: "1.36")]
   public class QueryRooms : QuerySpatialElements
   {
     public override Guid ComponentGuid => new Guid("5DDCB816-61A3-480F-AC45-67F66BEB2E78");
@@ -451,8 +424,7 @@ namespace RhinoInside.Revit.GH.Components.Topology
 
     protected override void TrySolveInstance(IGH_DataAccess DA)
     {
-      if (!Parameters.Document.TryGetDocumentOrCurrent(this, DA, "Document", out var doc)) return;
-
+      if (!Params.TryGetData(DA, "Model", out Types.IGH_ModelInstance model, x => x.IsValid)) return;
       if (!Params.TryGetData(DA, "Point", out Point3d? point)) return;
       var xyz = point.HasValue ? point.Value.ToXYZ() : default;
       if (!Params.TryGetData(DA, "Placed", out bool? placed)) return;
@@ -461,7 +433,7 @@ namespace RhinoInside.Revit.GH.Components.Topology
       if (!Params.TryGetData(DA, "Level", out Types.Level level)) return;
       if (!Params.TryGetData(DA, "Phase", out Types.Phase phase)) return;
       if (phase is null && Params.IndexOfInputParam("Phase") < 0)
-        phase = new Types.Phase(doc.Value.Phases.Cast<ARDB.Phase>().LastOrDefault());
+        phase = new Types.Phase(model.ModelDocument.Value.Phases.Cast<ARDB.Phase>().LastOrDefault());
       if (!Params.TryGetData(DA, "Enclosed", out bool? enclosed)) return;
       if (!Params.TryGetData(DA, "Filter", out ARDB.ElementFilter filter)) return;
 
@@ -472,7 +444,7 @@ namespace RhinoInside.Revit.GH.Components.Topology
       }
 
       var tol = GeometryTolerance.Model;
-      using (var collector = new ARDB.FilteredElementCollector(doc.Value))
+      using (var collector = new ARDB.FilteredElementCollector(model.ModelDocument.Value))
       {
         var elementsCollector = collector.WherePasses(ElementFilter);
 
@@ -480,7 +452,7 @@ namespace RhinoInside.Revit.GH.Components.Topology
           elementsCollector = elementsCollector.WherePasses(new ARDB.BoundingBoxContainsPointFilter(xyz));
 
         if (filter is object)
-          elementsCollector = elementsCollector.WherePasses(filter);
+          elementsCollector = elementsCollector.WherePasses(filter, model.ModelInstance.Value);
 
         if (TryGetFilterStringParam(ARDB.BuiltInParameter.ROOM_NUMBER, ref number, out var numberFilter))
           elementsCollector = elementsCollector.WherePasses(numberFilter);
@@ -511,12 +483,12 @@ namespace RhinoInside.Revit.GH.Components.Topology
         if (xyz is object)
           rooms = rooms.Where(room => room.Value.IsPointInRoom(xyz));
 
-        DA.SetDataList("Rooms",rooms.TakeWhileIsNotEscapeKeyDown(this));
+        DA.SetDataList("Rooms",rooms.AtModel(model).TakeWhileIsNotEscapeKeyDown(this));
       }
     }
   }
 
-  [ComponentVersion(introduced: "1.7", updated: "1.9")]
+  [ComponentVersion(introduced: "1.7", updated: "1.36")]
   public class QuerySpaces : QuerySpatialElements
   {
     public override Guid ComponentGuid => new Guid("A1CCF034-AA1F-4731-9863-3C22E0644E2B");
@@ -552,8 +524,7 @@ namespace RhinoInside.Revit.GH.Components.Topology
 
     protected override void TrySolveInstance(IGH_DataAccess DA)
     {
-      if (!Parameters.Document.TryGetDocumentOrCurrent(this, DA, "Document", out var doc)) return;
-
+      if (!Params.TryGetData(DA, "Model", out Types.IGH_ModelInstance model, x => x.IsValid)) return;
       if (!Params.TryGetData(DA, "Point", out Point3d? point)) return;
       var xyz = point.HasValue ? point.Value.ToXYZ() : default;
       if (!Params.TryGetData(DA, "Placed", out bool? placed)) return;
@@ -562,7 +533,7 @@ namespace RhinoInside.Revit.GH.Components.Topology
       if (!Params.TryGetData(DA, "Level", out Types.Level level)) return;
       if (!Params.TryGetData(DA, "Phase", out Types.Phase phase)) return;
       if (phase is null && Params.IndexOfInputParam("Phase") < 0)
-        phase = new Types.Phase(doc.Value.Phases.Cast<ARDB.Phase>().LastOrDefault());
+        phase = new Types.Phase(model.ModelDocument.Value.Phases.Cast<ARDB.Phase>().LastOrDefault());
       if (!Params.TryGetData(DA, "Enclosed", out bool? enclosed)) return;
       if (!Params.TryGetData(DA, "Filter", out ARDB.ElementFilter filter)) return;
 
@@ -573,7 +544,7 @@ namespace RhinoInside.Revit.GH.Components.Topology
       }
 
       var tol = GeometryTolerance.Model;
-      using (var collector = new ARDB.FilteredElementCollector(doc.Value))
+      using (var collector = new ARDB.FilteredElementCollector(model.ModelDocument.Value))
       {
         var elementsCollector = collector.WherePasses(ElementFilter);
 
@@ -581,7 +552,7 @@ namespace RhinoInside.Revit.GH.Components.Topology
           elementsCollector = elementsCollector.WherePasses(new ARDB.BoundingBoxContainsPointFilter(xyz));
 
         if (filter is object)
-          elementsCollector = elementsCollector.WherePasses(filter);
+          elementsCollector = elementsCollector.WherePasses(filter, model.ModelInstance.Value);
 
         if (TryGetFilterStringParam(ARDB.BuiltInParameter.ROOM_NUMBER, ref number, out var numberFilter))
           elementsCollector = elementsCollector.WherePasses(numberFilter);
@@ -612,7 +583,7 @@ namespace RhinoInside.Revit.GH.Components.Topology
         if (xyz is object)
           spaces = spaces.Where(room => room.Value.IsPointInSpace(xyz));
 
-        DA.SetDataList("Spaces", spaces.TakeWhileIsNotEscapeKeyDown(this));
+        DA.SetDataList("Spaces", spaces.AtModel(model).TakeWhileIsNotEscapeKeyDown(this));
       }
     }
   }