+++ src/RhinoInside.Revit.GH/Types/Instances/RevitLinkInstance.cs
@@ -1,13 +1,26 @@
 using System;
+using Grasshopper.Kernel.Types;
 using ARDB = Autodesk.Revit.DB;
 
 namespace RhinoInside.Revit.GH.Types
 {
+  [Kernel.Attributes.Name("Model")]
+  public interface IGH_ModelInstance : IGH_Goo
+  {
+    public RevitLinkInstance ModelInstance { get; }
+    public Document ModelDocument { get; }
+  }
+
   [Kernel.Attributes.Name("Linked Model")]
-  public sealed class RevitLinkInstance : Instance
+  public sealed class RevitLinkInstance : Instance, IGH_ModelInstance
   {
     protected override Type ValueType => typeof(ARDB.RevitLinkInstance);
     public new ARDB.RevitLinkInstance Value => base.Value as ARDB.RevitLinkInstance;
+
+    #region IGH_ModelInstance
+    RevitLinkInstance IGH_ModelInstance.ModelInstance => this;
+    public Document ModelDocument => Types.Document.FromValue(Value.GetLinkDocument());
+    #endregion
 
     public RevitLinkInstance() { }
     public RevitLinkInstance(ARDB.RevitLinkInstance instance) : base(instance) { }