+++ docs/pages/_en/1.0/reference/release-notes.md
@@ -10,6 +10,7 @@ group: Deployment & Configs
 
 ### WIP
 
+- Added support for linked elements on query components.
 - Added 'Query References' component.
 
 ### RC